1 to 25 projects/14-Photo Manipulation/requirements.txt
numpy>=1.23.0
Pillow>=10.0.0


scipy==1.10.1
tqdm==4.65.0